+++ mobile/android/app/src/main/kotlin/com/morethanhuman/curitalk/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.morethanhuman.curitalk
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()